requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-jose
passlib[bcrypt]
python-dotenv
